requirements.txt
click
dacite
jinja2
psutil
pyyaml
tabulate
